chore(paypal-connector): rename instances of mock to paypal
