+++ enabler/src/payment-enabler/payment-enabler-paypal.ts
@@ -11,11 +11,11 @@ declare global {
   }
 }
 
-export class MockPaymentEnabler implements PaymentEnabler {
+export class PaypalPaymentEnabler implements PaymentEnabler {
   setupData: Promise<{ baseOptions: BaseOptions }>;
 
   private constructor(options: EnablerOptions) {
-    this.setupData = MockPaymentEnabler._Setup(options);
+    this.setupData = PaypalPaymentEnabler._Setup(options);
   }
 
   private static _Setup = async (options: EnablerOptions): Promise<{ baseOptions: BaseOptions }> => {
