+++ processor/src/processors/paypal-operation.processor.ts
@@ -1,7 +1,7 @@
 import { healthCheckCommercetoolsPermissions, statusHandler } from '@commercetools/connect-payments-sdk';
-import { config } from '../../config/config';
-import { PaymentModificationStatus } from '../../dtos/operations/payment-intents.dto';
-import { paymentSDK } from '../../payment-sdk';
+import { config } from '../config/config';
+import { PaymentModificationStatus } from '../dtos/operations/payment-intents.dto';
+import { paymentSDK } from '../payment-sdk';
 import {
   CancelPaymentRequest,
   CapturePaymentRequest,
@@ -9,11 +9,11 @@ import {
   PaymentProviderModificationResponse,
   RefundPaymentRequest,
   StatusResponse,
-} from '../types/operation.type';
-import { OperationProcessor } from './operation.processor';
-const packageJSON = require('../../../package.json');
+} from '../services/types/operation.type';
+import { OperationProcessor } from '../services/processors/operation.processor';
+const packageJSON = require('../../package.json');
 
-export class MockOperationProcessor implements OperationProcessor {
+export class PaypalOperationProcessor implements OperationProcessor {
   async config(): Promise<ConfigResponse> {
     return {
       clientKey: config.mockClientKey,
