+++ processor/src/server/plugins/paypal-payment.plugin.ts
@@ -1,16 +1,16 @@
 import { FastifyInstance } from 'fastify';
 import { paymentSDK } from '../../payment-sdk';
-import { paymentRoutes } from '../../routes/mock-payment.route';
-import { MockPaymentService } from '../../services/mock-payment.service';
+import { paymentRoutes } from '../../routes/paypal-payment.route';
+import { PaypalPaymentService } from '../../services/paypal-payment.service';
 
 export default async function (server: FastifyInstance) {
-  const mockPaymentService = new MockPaymentService({
+  const paypalPaymentService = new PaypalPaymentService({
     ctCartService: paymentSDK.ctCartService,
     ctPaymentService: paymentSDK.ctPaymentService,
   });
 
   await server.register(paymentRoutes, {
-    paymentService: mockPaymentService,
+    paymentService: paypalPaymentService,
     sessionAuthHook: paymentSDK.sessionAuthHookFn,
   });
 }
